Move DQMGenericClient and QualityTester
... to DQMServices/Components/plugins, with the goal of eliminating the
ClientConfig package entirely.

Also merge the headers into .cc and remove the ROOT5 support code in
DQMGenericClient.
