--- DQMServices/ClientConfig/plugins/SealModule.cc
@@ -1,5 +1,0 @@
-#include "FWCore/Framework/interface/MakerMacros.h"
-
-#include "DQMServices/ClientConfig/interface/DQMGenericClient.h"
-
-DEFINE_FWK_MODULE(DQMGenericClient);
