+++ DQMServices/Components/plugins/DQMGenericClient.cc
@@ -1,27 +1,148 @@
 /*
  *  Class:DQMGenericClient 
  *
+ *  DQM histogram post processor
  *
  * 
  *  \author Junghwan Goh - SungKyunKwan University
  */
 
-#include "DQMServices/ClientConfig/interface/DQMGenericClient.h"
-
-#include "DQMServices/ClientConfig/interface/FitSlicesYTool.h"
-#include "DQMServices/Core/interface/DQMStore.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "DQMServices/Core/interface/DQMEDHarvester.h"
+#include "DQMServices/ClientConfig/interface/FitSlicesYTool.h"
 
+#include <TH1.h>
 #include <TH1F.h>
 #include <TClass.h>
 #include <TString.h>
 #include <TPRegexp.h>
+#include <TEfficiency.h>
 
+#include <set>
 #include <cmath>
+#include <string>
+#include <vector>
 #include <climits>
 #include <boost/tokenizer.hpp>
+
+
+class DQMGenericClient : public DQMEDHarvester {
+public:
+  DQMGenericClient(const edm::ParameterSet& pset);
+  ~DQMGenericClient() override{};
+
+  void dqmEndLuminosityBlock(DQMStore::IBooker& ibooker,
+                             DQMStore::IGetter& igetter,
+                             const edm::LuminosityBlock& lumiSeg,
+                             const edm::EventSetup& c) override;
+  void dqmEndJob(DQMStore::IBooker&, DQMStore::IGetter&) override;
+
+  enum class EfficType { none = 0, efficiency, fakerate, simpleratio };
+
+  struct EfficOption {
+    std::string name, title;
+    std::string numerator, denominator;
+    EfficType type;
+    bool isProfile;
+  };
+
+  struct ResolOption {
+    std::string namePrefix, titlePrefix;
+    std::string srcName;
+  };
+
+  struct ProfileOption {
+    std::string name, title;
+    std::string srcName;
+  };
+
+  struct NormOption {
+    std::string name, normHistName;
+  };
+
+  struct CDOption {
+    std::string name;
+    bool ascending;
+  };
+
+  struct NoFlowOption {
+    std::string name;
+  };
+
+  void computeEfficiency(DQMStore::IBooker& ibooker,
+                         DQMStore::IGetter& igetter,
+                         const std::string& startDir,
+                         const std::string& efficMEName,
+                         const std::string& efficMETitle,
+                         const std::string& recoMEName,
+                         const std::string& simMEName,
+                         const EfficType type = EfficType::efficiency,
+                         const bool makeProfile = false);
+  void computeResolution(DQMStore::IBooker& ibooker,
+                         DQMStore::IGetter& igetter,
+                         const std::string& startDir,
+                         const std::string& fitMEPrefix,
+                         const std::string& fitMETitlePrefix,
+                         const std::string& srcMEName);
+  void computeProfile(DQMStore::IBooker& ibooker,
+                      DQMStore::IGetter& igetter,
+                      const std::string& startDir,
+                      const std::string& profileMEName,
+                      const std::string& profileMETitle,
+                      const std::string& srcMEName);
+
+  void normalizeToEntries(DQMStore::IBooker& ibooker,
+                          DQMStore::IGetter& igetter,
+                          const std::string& startDir,
+                          const std::string& histName,
+                          const std::string& normHistName);
+  void makeCumulativeDist(DQMStore::IBooker& ibooker,
+                          DQMStore::IGetter& igetter,
+                          const std::string& startDir,
+                          const std::string& cdName,
+                          bool ascending = true);
+  void makeNoFlowDist(DQMStore::IBooker& ibooker,
+                      DQMStore::IGetter& igetter,
+                      const std::string& startDir,
+                      const std::string& cdName);
+
+  void limitedFit(MonitorElement* srcME, MonitorElement* meanME, MonitorElement* sigmaME);
+
+private:
+  unsigned int verbose_;
+  bool runOnEndLumi_;
+  bool runOnEndJob_;
+  bool makeGlobalEffPlot_;
+  bool isWildcardUsed_;
+  bool resLimitedFit_;
+
+  DQMStore* theDQM;
+  std::vector<std::string> subDirs_;
+  std::string outputFileName_;
+
+  std::vector<EfficOption> efficOptions_;
+  std::vector<ResolOption> resolOptions_;
+  std::vector<ProfileOption> profileOptions_;
+  std::vector<NormOption> normOptions_;
+  std::vector<CDOption> cdOptions_;
+  std::vector<NoFlowOption> noFlowOptions_;
+
+  void generic_eff(TH1* denom, TH1* numer, MonitorElement* efficiencyHist, const EfficType type = EfficType::efficiency);
+
+  void findAllSubdirectories(DQMStore::IBooker& ibooker,
+                             DQMStore::IGetter& igetter,
+                             std::string dir,
+                             std::set<std::string>* myList,
+                             const TString& pattern);
+
+  void makeAllPlots(DQMStore::IBooker&, DQMStore::IGetter&);
+
+  void removeMEIfBooked(const std::string& meName, DQMStore::IGetter& igetter);
+
+};
 
 using namespace std;
 using namespace edm;
@@ -1062,5 +1183,7 @@ void DQMGenericClient::generic_eff(TH1* denom, TH1* numer, MonitorElement* effic
   //efficiencyHist->setMinimum(0.0);
   //efficiencyHist->setMaximum(1.0);
 }
+
+DEFINE_FWK_MODULE(DQMGenericClient);
 
 /* vim:set ts=2 sts=2 sw=2 expandtab: */
